requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
dnspython==2.4.2
duckdb==0.10.2
idna==3.6
numpy==1.26.3
pandas==2.1.4
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tqdm==4.66.1
tzdata==2023.4
urllib3==2.1.0
apache-airflow==2.8.1
redis==4.1.1
nltk==3.6.5
aiohttp
aiofiles
